third_party/blink/web_tests/external/wpt/html/browsers/history/the-history-interface/history_pushstate_url_rewriting-expected.txt

This is a testharness.js-based test.
[FAIL] data:(script to run this test) to data:(script to run this test)?newsearch should not work
  assert_equals: expected "exception" but got "no exception"
Harness: the test ran to completion.
